+++ javaScript/projects/weather-app/src/components/tabs/forecast/forecast.config.js
@@ -8,40 +8,10 @@
 // ENE, E, ESE
 // SE, SSE, S, SSW, SW
 // WSW, W, WNW
-import importAll from '../../../helpers/importAll';
+// import importAll from '../../../helpers/importAll';
 import formatTime from '../../../helpers/formatTime';
 import formatDate from '../../../helpers/formatDate';
-
-// const icons = importAll(require.context('../../../assets/icons', false, /\.svg$/));
-const unitSystems = {
-  icons: importAll(require.context('../../../assets/icons', false, /\.svg$/)),
-  metric: {
-    temp: 'c',
-    speed: 'kph',
-    precipitation: 'mm',
-    pressure: 'mb',
-    distance: 'km',
-  },
-  imperial: {
-    temp: 'f',
-    speed: 'mph',
-    precipitation: 'in',
-    pressure: 'in',
-    distance: 'miles',
-  },
-  get(key) {
-    return this.unitSystem[key];
-  },
-  setIcon(key) {
-    return this.icons.files[Object.keys(this.icons.files).find((iconKey) => iconKey.includes(key))];
-  },
-  roundValue(value) {
-    return Math.round(value);
-  },
-  setValue(obj, ...args) {
-    return this.roundValue(obj[`${args[0]}${this.get(args[1])}`]);
-  },
-};
+import unitSystems from '../unitsystems';
 
 const data = (state) => [
   {